init.target.rc: qrngd: 8916: Start a qrng post boot executable.
QRNG bus scale is set to high before qrng_read and set to low after
qrng_read, this impacts the performance on boot-up, as there are
various clients using qrng during bootup.Inorder to fix this problem,
an executable is introduced which calls into kernel module (that sets
low bandwidth). This executable is started once the boot is complete
and this taps onto the kernel module (thought IOCTL) to reset the
bus_bandwidth.

Change-Id: Iec9b191545d05d15c04edaabd14951029485d16c
Signed-off-by: AnilKumar Chimata <anilc@codeaurora.org>