msm8916_64: added sbc, ml-listener and enterprise security jars
This commit required in order to enable sbc (security-boot-check)
and ml-listener to run automatically at boot time.
The ml-listener calculates on-demands hashes of requested files and
partitions. In order to do that, it needs to run with root privilege.
The calculated hashes are being saved in the SFS,
and can only be read by an authenticated request.
In addition, this commit required in order to enable a qsb, one of the SEAL
(Snapdragon Enterprise Access Lock) features, to use a supplied port
by adding the appropriate jars to the msm8916_64.mk file.

Change-Id: I3f0653d39ce371f70eabdde7cf0cb37234482d4f